requirements.txt
python-dotenv==1.0.1
openai==1.16.2
Wikidata==0.7.0